fix: delete throws RuntimeException("Doctor not found");

+++ src/main/java/ru/shmelev/stomatologyapp/service/DoctorServiceImpl.java
@@ -40,6 +40,9 @@ public class DoctorServiceImpl implements DoctorService {
 
     @Override
     public void delete(Long id) {
+        if (doctorRepository.existsById(id)==false) {
+            throw new RuntimeException("Doctor not found");
+        }
         doctorRepository.deleteById(id);
     }
